+++ src/main/java/org/wikicrimes/util/kernelmap/renderer/KMFunctionBender.java
@@ -8,30 +8,30 @@ import java.awt.Color;
  * Distorce a função de densidade do "originalRenderer" pra aumentar ou diminuir 
  * a diferença entre as densidades altas e baixas.
  * 
- * Para "bending = 0", não há distorção
- * Para "bending > 0", as densidades baixas são enfatizadas  
- * Para "bending < 0", as densidades altas são enfatizadas
+ * Para "bending = 0", não ha distorção
+ * Para "bending > 0", as densidades baixas sao enfatizadas  
+ * Para "bending < 0", as densidades altas sao enfatizadas
  */
 public class KMFunctionBender extends CellBasedKMR{
 
 	
 	private CellBasedKMR originalRenderer;
-	private double bending;
+	private float bending;
 	
-	public KMFunctionBender(CellBasedKMR renderer, double bending){
+	public KMFunctionBender(CellBasedKMR renderer, float bending){
 		super(renderer.kernel);
 		this.originalRenderer = renderer;
 		this.bending = bending;
 	}
 	
 	@Override
-	public Color renderCell(double density) {
-		double newDensity = bend(density);
+	public Color renderCell(float density) {
+		float newDensity = bend(density);
 		return originalRenderer.renderCell(newDensity);
 	}
 	
-	private double bend(double density){
-		return Math.pow(density/MAX_DENSITY, bending)*MAX_DENSITY;
+	private float bend(float density){
+		return (float)Math.pow(density/MAX_DENSITY, bending)*MAX_DENSITY;
 	}
 	
 }
