+++ src/main/java/org/wikicrimes/util/kernelmap/renderer/RainbowKMR.java
@@ -14,33 +14,31 @@ public class RainbowKMR extends CellBasedKMR{
 		super(kernel);
 	}
 	
-	public RainbowKMR(KernelMap kernel, double maxDensity){
+	public RainbowKMR(KernelMap kernel, float maxDensity){
 		super(kernel, maxDensity);
 	}
 	
-	public Color renderCell(double densidade){
+	public Color renderCell(float densidade){
 		if(densidade == 0)
 			return new Color(1,1,1,0);
 		final float alpha = .5f;
-		double quarto = MAX_DENSITY/4;
+		float quarto = MAX_DENSITY/4;
 		if(densidade <= quarto){
-			float g = (float)(densidade/quarto);
+			float g = densidade/quarto;
 			return new Color(0,g,1,alpha);
 		}else if(densidade <= 2*quarto){
 			densidade -= quarto;
-			float b = (float)(1 - densidade/quarto);
+			float b = 1 - densidade/quarto;
 			return new Color(0,1,b,alpha);
 		}else if(densidade <= 3*quarto){
 			densidade -= 2*quarto;
-			float r = (float)(densidade/quarto);
+			float r = densidade/quarto;
 			return new Color(r,1,0,alpha);
 		}else{
 			densidade -= 3*quarto;
-			float g = (float)(1 - densidade/quarto);
+			float g = 1 - densidade/quarto;
 			return new Color(1,g,0,alpha);
 		}
 	}
-	
-	
-	
+		
 }
