+++ src/main/java/org/wikicrimes/util/kernelmap/renderer/TransparentToColorKMR.java
@@ -16,12 +16,12 @@ public class TransparentToColorKMR extends CellBasedKMR{
 		this.color = color;
 	}
 	
-	public TransparentToColorKMR(KernelMap kernel, Color color, double maxDensity){
+	public TransparentToColorKMR(KernelMap kernel, Color color, float maxDensity){
 		super(kernel, maxDensity);
 		this.color = color;
 	}
 	
-	public Color renderCell(double densidade){
+	public Color renderCell(float densidade){
 		float alpha = (float)(densidade/MAX_DENSITY);
 		if(alpha<0) alpha = 0;
 		if(alpha>1) alpha = 1;
